Short term check in as a work around for [Bug 7669] Unable to recieve mail on Ubuntu. Something is going on with schema.Long where the default value is not being persisted to the mail (twisted) thread. Changing to an Integer seems to fix the problem. However, the value of IMAP UID should be a long. So once the problem is resolved the messageDownloadSequence will be changed back to schema.Long
git-svn-id: http://svn.osafoundation.org/chandler/trunk@12802 3c49585b-f0f7-0310-b5f9-dfe92a88fbfe

+++ chandler/parcels/osaf/mail/imap.py
@@ -286,7 +286,7 @@ class IMAPClient(base.AbstractDownloadClient):
         nextUID = self._getNextUID()
 
         for message in msgs.itervalues():
-            luid = long(message['UID'])
+            luid = int(message['UID'])
 
             if luid < nextUID:
                 continue
